+++ src/main/java/com/example/picket/domain/auth/service/OAuthService.java
@@ -6,58 +6,29 @@ import com.example.picket.common.enums.Gender;
 import com.example.picket.common.enums.OAuth;
 import com.example.picket.common.enums.UserRole;
 import com.example.picket.config.GoogleOAuthProperties;
+import com.example.picket.domain.auth.client.GoogleOauthClient;
 import com.example.picket.domain.user.entity.User;
 import com.example.picket.domain.user.repository.UserRepository;
 import com.example.picket.domain.user.service.UserQueryService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.reactive.function.client.WebClient;
 
 import java.time.LocalDate;
-import java.util.Map;
 
 @RequiredArgsConstructor
 @Service
 public class OAuthService {
 
-    private final GoogleOAuthProperties googleOAuthProperties;
-    private final WebClient webClient;
     private final UserQueryService userQueryService;
     private final UserRepository userRepository;
-
-    public String getAccessToken(String code, String redirectUrl) {
-        return webClient.post()
-                .uri(googleOAuthProperties.getTokenUri())
-                .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-                .bodyValue("code="+code
-                        +"&client_id="+googleOAuthProperties.getClientId()
-                        +"&client_secret="+googleOAuthProperties.getClientSecret()
-                        +"&redirect_uri="+redirectUrl
-                        +"&grant_type=authorization_code")
-                .retrieve()
-                .bodyToMono(Map.class)
-                .map(response -> (String)response.get("access_token"))
-                .block();
-    }
-
-    public OauthUser getUser(String accessToken){
-        return webClient.get()
-                .uri(googleOAuthProperties.getUserInfoUri())
-                .headers(headers -> headers.setBearerAuth(accessToken))
-                .retrieve()
-                .bodyToMono(OauthUser.class)
-                .block();
-    }
-
+    private final GoogleOauthClient googleOauthClient;
 
     @Transactional
     public User getOrCreateUser(HttpSession session, String code, UserRole userRole) {
-        String accessToken = getAccessToken(code, googleOAuthProperties.getRedirectUriByRole(userRole));
-        OauthUser oauthUser = getUser(accessToken);
+        String accessToken = googleOauthClient.getAccessToken(code, userRole);
+        OauthUser oauthUser = googleOauthClient.getUser(accessToken);
         User user = userQueryService.getUserByEmail(oauthUser.getEmail())
                 .orElseGet( () -> userRepository.save(
                         User.toOAuthEntity(oauthUser.getEmail(),
@@ -79,9 +50,9 @@ public class OAuthService {
     }
 
     @Transactional
-    public void signup(Long id, String nickname, LocalDate birth, Gender gender, UserRole userRole) {
+    public void signup(Long id, String nickname, LocalDate birth, Gender gender) {
         User user = userQueryService.getUser(id);
-        user.oAuthSignup(userRole, null, nickname, birth, gender);
+        user.oAuthSignup(nickname, birth, gender);
     }
 
 }
